requirements.txt
streamlit
python-dotenv
Pillow
pdf2image
google-generativeai
reportlab